Exception handling for playback, taskbar, and move DEA to Stat Analysis
