+++ src/edu/iastate/metnet/metaomgraph/ui/TaskbarPanel.java
@@ -78,24 +78,24 @@ public class TaskbarPanel extends JPanel{
 
 	private LinkedHashMap<String,List<TaskbarInternalFrame>> taskbarData;
 	private JMenuBar menuBar;
-	
+
 	/**
 	 * Constructor to initialize the masterdata (taskbarData) and the JPanel
 	 */
 	public TaskbarPanel() {
-		
+
 		GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment();
 		GraphicsDevice defaultScreen = ge.getDefaultScreenDevice();
 		Rectangle rect = defaultScreen.getDefaultConfiguration().getBounds();
 		setSize((int)rect.getMaxX(), 30);
 		setLocation(0, (int)rect.getMaxY()-30);
 		setLayout(new FlowLayout(FlowLayout.LEFT));
-		
+
 		taskbarData = new LinkedHashMap<String,List<TaskbarInternalFrame>>();
 		menuBar = new JMenuBar();
 		add(menuBar);
 	}
-	
+
 	/**
 	 * This method adds a tab to the taskbar
 	 * 
@@ -110,7 +110,7 @@ public class TaskbarPanel extends JPanel{
 	 * 
 	 */
 	public void addToTaskbar(TaskbarInternalFrame currentFrame) {
-		
+
 		if(taskbarData.get(currentFrame.getModel().getFrameType()) != null) {
 			ArrayList<TaskbarInternalFrame> listOfFrames = (ArrayList<TaskbarInternalFrame>)taskbarData.get(currentFrame.getModel().getFrameType());
 			listOfFrames.add(currentFrame);
@@ -120,13 +120,13 @@ public class TaskbarPanel extends JPanel{
 			listOfFrames.add(currentFrame);
 			taskbarData.put(currentFrame.getModel().getFrameType(), listOfFrames);
 		}
-		
+
 		reloadTaskbar();
-		
+
 	}
-	
-	
-	
+
+
+
 	/**
 	 * This method removes the given frame from the taskbar.
 	 * 
@@ -140,11 +140,11 @@ public class TaskbarPanel extends JPanel{
 	 * At the end, the taskbar is reloaded to reflect the removal.
 	 */
 	public void removeFromTaskbar(TaskbarInternalFrame currentFrame) {
-		
+
 		if(taskbarData.get(currentFrame.getModel().getFrameType()) != null) {
 			ArrayList<TaskbarInternalFrame> listOfFrames = (ArrayList<TaskbarInternalFrame>)taskbarData.get(currentFrame.getModel().getFrameType());
 			listOfFrames.remove(currentFrame);
-			
+
 			if(listOfFrames.size()==0) {
 				taskbarData.remove(currentFrame.getModel().getFrameType());
 			}
@@ -152,23 +152,23 @@ public class TaskbarPanel extends JPanel{
 		else {
 			taskbarData.remove(currentFrame.getModel().getFrameType());
 		}
-		
+
 		reloadTaskbar();
-		
+
 	}
 
-	
+
 	/**
 	 * Clears the tab data from the taskbarData master data, and removes all the
 	 * JMenu objects from the menubar.
 	 */
 	public void removeAllTabsFromTaskbar() {
-		
+
 		taskbarData.clear();
 		menuBar.removeAll();
 	}
-	
-	
+
+
 	/**
 	 * This method first removes all the JMenus from the menuBar object.
 	 * 
@@ -181,22 +181,22 @@ public class TaskbarPanel extends JPanel{
 	 *
 	 */
 	public void reloadTaskbar() {
-		
+
 		menuBar.removeAll();
-		
+
 		if(taskbarData != null) {
-			
+
 			for(Map.Entry<String, List<TaskbarInternalFrame>> entryset : taskbarData.entrySet()) {
 				JMenu menu = new JMenu(entryset.getKey());
 				List<TaskbarInternalFrame> typeFrames = entryset.getValue();
-				
+
 				if(typeFrames != null) {
 					for(TaskbarInternalFrame frame : typeFrames) {
 						JMenuItem menuItem = new JMenuItem(frame.getModel().getFrameName());
 						JPanel tabItemPanel = new JPanel();
 						tabItemPanel.setLayout(new BorderLayout());
 						JLabel tabItemLabel = new JLabel(frame.getModel().getFrameName());
-						
+
 						tabItemPanel.add(tabItemLabel);
 
 						UIManager.put("MenuItem.selectionForeground", Color.BLUE);
@@ -204,25 +204,32 @@ public class TaskbarPanel extends JPanel{
 						menu.add(new JSeparator());
 						menu.add(menuItem);
 						menu.setMargin(new Insets(2, 8, 2, 8));
-				
-						menuItem.addActionListener(new java.awt.event.ActionListener() {
-						    public void actionPerformed(java.awt.event.ActionEvent evt) {
 
-						    	frame.getDesktopPane().getDesktopManager().maximizeFrame(frame);
-						    	frame.getDesktopPane().getDesktopManager().minimizeFrame(frame);
-						    	frame.moveToFront();
-						    }
+						menuItem.addActionListener(new java.awt.event.ActionListener() {
+							public void actionPerformed(java.awt.event.ActionEvent evt) {
+
+								try {
+									if(frame != null) {
+										frame.getDesktopPane().getDesktopManager().maximizeFrame(frame);
+										frame.getDesktopPane().getDesktopManager().minimizeFrame(frame);
+										frame.moveToFront();
+									}
+								}
+								catch(Exception e) {
+
+								}
+							}
 						});
 					}
 				}
-				
+
 				JSeparator menuSeparator = new JSeparator();
 				menuBar.add(menuSeparator);
 				menuBar.add(menu);
-				
+
 			}
 		}
-	
+
 	}
-	
+
 }
